+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Improvements:
 
 ## Bug fixes:
+* Fix: AfterScenario hook is not called when a scenario end in inconclusive or skipped status (#883, #900)
 * Fix: Feature-level tags are not generating MSTest TestCategory attributes for scenarios (#910)
 
 *Contributors of this release (in alphabetical order):* @clrudolphi